auth/pyproject.toml
[build-system]  # https://python-poetry.org/docs/pyproject/#poetry-and-pep-517
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]  # https://python-poetry.org/docs/pyproject/
package-mode = false
name = "auth"

[tool.poetry.dependencies]  # https://python-poetry.org/docs/dependency-specification/
python = "^3.12"
pydantic = "^2.6.4"
fastapi = "^0.110.1"
async-fastapi-jwt-auth = "^0.6.4"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.29" }
redis = { extras = ["hiredis"], version = "^5.0.3" }
uvicorn = { extras = ["standart"], version = "^0.29.0" }
gunicorn = "^21.2.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
werkzeug = "^3.0.2"
orjson = "^3.10.1"
typer-slim = "^0.12.3"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-jaeger = "^1.21.0"
opentelemetry-instrumentation-sqlalchemy = "^0.45b0"

[tool.poetry.group.dev.dependencies]  # https://python-poetry.org/docs/master/managing-dependencies/
mypy = "^1.9.0"
pre-commit = "^3.7.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
coverage = "^7.4.4"
typeguard = "^4.2.1"
ruff = "^0.3.4"
safety = "^3.1.0"
deptry = "^0.16.0"
types-redis = "^4.6.0"
testcontainers = { extras = ["postgres", "redis"], version = "^4.4.0" }
pytest-asyncio = "^0.23.6"
pytest-unordered = "^0.6.0"
pytest-alembic = "^0.11.0"
pytest-postgresql = "^6.0.0"
pip = "^24.0"

[tool.mypy]  # https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
junit_xml = "reports/mypy.xml"
python_version = "3.12"
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
show_traceback = true
color_output = true
plugins = ["pydantic.mypy"]

strict = false  # https://mypy.readthedocs.io/en/stable/existing_code.html#introduce-stricter-options
ignore_missing_imports = true
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false

[tool.pydantic-mypy]  # https://pydantic-docs.helpmanual.io/mypy_plugin/#configuring-the-plugin
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
filterwarnings = ["error", "ignore::DeprecationWarning", "ignore::ImportWarning", "ignore::ResourceWarning"]
testpaths = ["auth", "tests"]
asyncio_mode = "auto"
xfail_strict = true

# Extra options:
addopts = [
  "--color=yes",
  "--failed-first",
  "--strict-config",
  "--strict-markers",
  "--tb=short",
  "--verbosity=2",
  "--doctest-modules",
  "--doctest-continue-on-failure",
  "--cov",
  "--junitxml=reports/junit.xml",
  "--test-alembic", # pytest-alembic
]

[tool.coverage.run]  # https://coverage.readthedocs.io/en/latest/config.html#run
branch = true
data_file = "reports/.coverage"
source = ["auth"]

[tool.coverage.report]  # https://coverage.readthedocs.io/en/latest/config.html#report
precision = 1
show_missing = true
skip_covered = true

[tool.coverage.xml]  # https://coverage.readthedocs.io/en/latest/config.html#xml
output = "reports/coverage.xml"

[tool.pylint.basic]
class-attribute-naming-style = "any"
class-const-naming-style = "any"
good-names = ["i", "j", "k", "e", "x", "id", "_", "pk"]

[tool.pylint.exceptions]
overgeneral-exceptions = ["builtins.BaseException", "builtins.Exception"]

[tool.ruff]  # https://github.com/charliermarsh/ruff
line-length = 100
target-version = "py312"
exclude = ["migrations"]

[tool.ruff.lint]
ignore-init-module-imports = true
select = ["ALL"]
ignore = [
  "TD", # flake8-todos
  "FIX", # flake8-fixme
  "ANN", # flake8-annotations (handled by mypy)
  "A003", # Class attribute `id` is shadowing a Python builtin
  "EM", # flake8-errmsg - String literals in exceptions
  "E501", # pycodestyle - line too long (handled by formatter)
  "D100", # pydocstyle - Missing docstring in public module
  "D101", # pydocstyle - Missing docstring in public class
  "D102", # pydocstyle - Missing docstring in public method
  "D103", # pydocstyle - Missing docstring in public function
  "D104", # pydocstyle - Missing docstring in public package
  "D105", # pydocstyle - Missing docstring in magic method
  "D106", # pydocstyle - Missing docstring in public nested class
  "D107", # pydocstyle - Missing docstring in __init__
  "COM812", # forced by ruff formatter
  "ISC001", # forced by ruff formatter
]
unfixable = [
  "F", # pyflakes
  "ERA", # eradicate - commented-out code
]
allowed-confusables = [
  "а", "А", "б", "Б", "в", "В", "г", "Г", "е", "Е", "з", "З", "и", "И", "к", "К", "м", "М",
  "н", "Н", "о", "О", "р", "Р", "с", "С", "у", "У", "ф", "Ф", "х", "Х"  # Allow cyrillic letters
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
  "F401", # Unused import
  "F403", # Import star
]
"tests/**/*.py" = [
  "S101", # Use of assert detected (assert allowed in tests)
  "PLR2004", # Magic value used in comparison
  "RUF001", # String contains ambiguous letter
]

[tool.ruff.lint.flake8-tidy-imports]  # https://docs.astral.sh/ruff/settings/#flake8-tidy-imports
ban-relative-imports = "all"

[tool.ruff.lint.isort]  # https://docs.astral.sh/ruff/settings/#isort
section-order = ["future", "typing", "standard-library", "third-party", "first-party", "local-folder"]
lines-between-types = 1
split-on-trailing-comma = false

[tool.ruff.lint.isort.sections]
"typing" = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]

[tool.ruff.lint.pycodestyle]  # https://docs.astral.sh/ruff/settings/#pycodestyle
max-doc-length = 100

[tool.ruff.lint.pydocstyle]  # https://docs.astral.sh/ruff/settings/#pydocstyle
convention = "google"

[tool.ruff.lint.pylint]  # https://docs.astral.sh/ruff/settings/#pylint
allow-magic-value-types = ["int", "str", "bytes"]
max-args = 10
max-statements = 30

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true

[tool.deptry.per_rule_ignores]
DEP002 = ['gunicorn', 'asyncpg', 'orjson']
DEP003 = []
